Finalize Light Dashboard Styles
I'm sure there will be more changes but I think it is good for now. It is mostly responsive and seems to work fine on all physical devices I have tested. The scrollbar in DevTools is showing up a little weird on mobile devices, but appears to be working on physical devices, but still might want to keep an eye on that.